Refactor[MQB]: move partition constants to mqbi
