+++ src/groups/mqb/mqbi/mqbi_dispatcher.h
@@ -567,10 +567,6 @@ class DispatcherConfirmEvent {
 
     /// Return whether this event is a relay event or not.
     virtual bool isRelay() const = 0;
-
-    /// Return the partitionId affected to the queue associated to this
-    /// confirm message.  This is only valid when `isRelay() == true`.
-    virtual int partitionId() const = 0;
 };
 
 // ===========================
@@ -706,10 +702,6 @@ class DispatcherPutEvent {
 
     /// Return whether this event is a relay event or not.
     virtual bool isRelay() const = 0;
-
-    /// Return the partitionId affected to the queue associated to this
-    /// put message.  This is only valid when `isRelay() == true`.
-    virtual int partitionId() const = 0;
 
     /// Return a reference not offering modifiable access to the put header
     /// associated to this event.  This protocol struct is only valid when
@@ -1271,34 +1263,6 @@ bsl::ostream& operator<<(bsl::ostream& stream, const DispatcherClient& client);
 // class DispatcherEvent
 // ---------------------
 
-inline DispatcherEvent::DispatcherEvent(bslma::Allocator* allocator)
-: d_type(DispatcherEventType::e_UNDEFINED)
-, d_source_p(0)
-, d_destination_p(0)
-, d_ackMessage()
-, d_blob_sp(0, allocator)
-, d_options_sp(0, allocator)
-, d_clusterNode_p(0)
-, d_confirmMessage()
-, d_rejectMessage()
-, d_controlMessage(allocator)
-, d_guid(bmqt::MessageGUID())
-, d_isRelay(false)
-, d_partitionId(-1)
-, d_putHeader()
-, d_queueHandle_p(0)
-, d_queueId(-1)
-, d_subQueueInfos(allocator)
-, d_messagePropertiesInfo()
-, d_compressionAlgorithmType(bmqt::CompressionAlgorithmType::e_NONE)
-, d_isOutOfOrder(false)
-, d_genCount(0)
-, d_callback(allocator)
-, d_finalizeCallback(allocator)
-{
-    // NOTHING
-}
-
 inline const bmqp::AckMessage& DispatcherEvent::ackMessage() const
 {
     return d_ackMessage;
@@ -1592,101 +1556,6 @@ DispatcherEvent::setState(const bsl::shared_ptr<bmqu::AtomicState>& state)
 {
     d_state = state;
     return *this;
-}
-
-inline void DispatcherEvent::reset()
-{
-    switch (d_type) {
-    case mqbi::DispatcherEventType::e_UNDEFINED: {
-        // NOTHING
-    } break;
-    case mqbi::DispatcherEventType::e_DISPATCHER: {
-        if (!d_finalizeCallback.empty()) {
-            // We only set finalizeCallback on e_DISPATCHER events
-
-            // TODO(678098): make a special event type that handles this case
-            d_finalizeCallback();
-        }
-
-        d_callback.reset();
-        d_finalizeCallback.reset();
-    } break;
-    case mqbi::DispatcherEventType::e_CALLBACK: {
-        d_callback.reset();
-    } break;
-    case mqbi::DispatcherEventType::e_CONTROL_MSG: {
-        d_controlMessage.reset();
-    } break;
-    case mqbi::DispatcherEventType::e_CONFIRM: {
-        d_blob_sp.reset();
-        d_clusterNode_p  = 0;
-        d_confirmMessage = bmqp::ConfirmMessage();
-        d_isRelay        = false;
-        d_partitionId    = -1;
-    } break;
-    case mqbi::DispatcherEventType::e_REJECT: {
-        d_blob_sp.reset();
-        d_clusterNode_p = 0;
-        d_rejectMessage = bmqp::RejectMessage();
-        d_isRelay       = false;
-        d_partitionId   = -1;
-    } break;
-    case mqbi::DispatcherEventType::e_PUSH: {
-        d_blob_sp.reset();
-        d_options_sp.reset();
-        d_clusterNode_p = 0;
-        d_guid          = bmqt::MessageGUID();
-        d_isRelay       = false;
-        d_queueId       = -1;
-        d_subQueueInfos.clear();
-        d_messagePropertiesInfo    = bmqp::MessagePropertiesInfo();
-        d_compressionAlgorithmType = bmqt::CompressionAlgorithmType::e_NONE;
-        d_isOutOfOrder             = false;
-    } break;
-    case mqbi::DispatcherEventType::e_PUT: {
-        d_blob_sp.reset();
-        d_options_sp.reset();
-        d_clusterNode_p = 0;
-        d_isRelay       = false;
-        d_partitionId   = -1;
-        d_putHeader     = bmqp::PutHeader();
-        d_queueHandle_p = 0;
-        d_genCount      = 0;
-        d_state.reset();
-    } break;
-    case mqbi::DispatcherEventType::e_ACK: {
-        d_ackMessage = bmqp::AckMessage();
-        d_blob_sp.reset();
-        d_options_sp.reset();
-        d_clusterNode_p = 0;
-        d_isRelay       = false;
-    } break;
-    case mqbi::DispatcherEventType::e_CLUSTER_STATE: {
-        d_blob_sp.reset();
-        d_clusterNode_p = 0;
-    } break;
-    case mqbi::DispatcherEventType::e_STORAGE: {
-        d_blob_sp.reset();
-        d_clusterNode_p = 0;
-        d_isRelay       = false;
-    } break;
-    case mqbi::DispatcherEventType::e_RECOVERY: {
-        d_blob_sp.reset();
-        d_clusterNode_p = 0;
-        d_isRelay       = false;
-    } break;
-    case mqbi::DispatcherEventType::e_REPLICATION_RECEIPT: {
-        d_blob_sp.reset();
-        d_clusterNode_p = 0;
-    } break;
-    default: {
-        BSLS_ASSERT_OPT(false && "Unexpected event type");
-    } break;
-    }
-
-    d_type          = DispatcherEventType::e_UNDEFINED;
-    d_source_p      = 0;
-    d_destination_p = 0;
 }
 
 inline DispatcherEventType::Enum DispatcherEvent::type() const
